🔥 Remove user info in profile service
