+++ app/src/main/java/com/mashup/mobalmobal/ui/profile/data/service/ProfileService.kt
@@ -2,7 +2,6 @@ package com.mashup.mobalmobal.ui.profile.data.service
 
 import com.mashup.mobalmobal.ui.profile.data.dto.MyDonateResponseDto
 import com.mashup.mobalmobal.ui.profile.data.dto.MyPostResponseDto
-import com.mashup.mobalmobal.ui.profile.data.dto.ProfileResponseDto
 import io.reactivex.Single
 import okhttp3.RequestBody
 import retrofit2.http.Body
@@ -10,9 +9,6 @@ import retrofit2.http.GET
 import retrofit2.http.POST
 
 interface ProfileService {
-    @GET("users")
-    fun getUserInfo(): Single<ProfileResponseDto>
-
     @POST("posts/my")
     fun getMyPosts(
         @Body
