+++ app/src/main/java/com/mashup/mobalmobal/ui/profile/presenter/ProfileViewModel.kt
@@ -11,6 +11,7 @@ import com.mashup.mobalmobal.data.repository.UserRepository
 import com.mashup.mobalmobal.ui.profile.data.dto.*
 import com.mashup.mobalmobal.ui.profile.data.repository.ProfileRepository
 import com.mashup.mobalmobal.ui.profile.domain.model.ProfileItem
+import com.mashup.mobalmobal.ui.profile.domain.model.toProfileItem
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.Observable
 import io.reactivex.rxkotlin.zipWith
